+++ websitebuild-20250429/src/Models/DeleteAppSceneRecipientRequest.php
@@ -1,0 +1,76 @@
+<?php
+
+// This file is auto-generated, don't edit it. Thanks.
+
+namespace AlibabaCloud\SDK\WebsiteBuild\V20250429\Models;
+
+use AlibabaCloud\Dara\Model;
+
+class DeleteAppSceneRecipientRequest extends Model
+{
+    /**
+     * @var string
+     */
+    public $bizId;
+
+    /**
+     * @var string
+     */
+    public $recipientId;
+
+    /**
+     * @var string
+     */
+    public $sceneId;
+    protected $_name = [
+        'bizId' => 'BizId',
+        'recipientId' => 'RecipientId',
+        'sceneId' => 'SceneId',
+    ];
+
+    public function validate()
+    {
+        parent::validate();
+    }
+
+    public function toArray($noStream = false)
+    {
+        $res = [];
+        if (null !== $this->bizId) {
+            $res['BizId'] = $this->bizId;
+        }
+
+        if (null !== $this->recipientId) {
+            $res['RecipientId'] = $this->recipientId;
+        }
+
+        if (null !== $this->sceneId) {
+            $res['SceneId'] = $this->sceneId;
+        }
+
+        return $res;
+    }
+
+    public function toMap($noStream = false)
+    {
+        return $this->toArray($noStream);
+    }
+
+    public static function fromMap($map = [])
+    {
+        $model = new self();
+        if (isset($map['BizId'])) {
+            $model->bizId = $map['BizId'];
+        }
+
+        if (isset($map['RecipientId'])) {
+            $model->recipientId = $map['RecipientId'];
+        }
+
+        if (isset($map['SceneId'])) {
+            $model->sceneId = $map['SceneId'];
+        }
+
+        return $model;
+    }
+}